+++ packages/js/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @trainheroic-unofficial/js
 
+## 0.4.2
+
+### Patch Changes
+
+- @trainheroic-unofficial/dto@0.4.2
+
 ## 0.4.1
 
 ### Patch Changes